Parallelizing #include parsing.
Returning an IAsyncEnumerable doesn't fetch all the results in the
background until you need them; instead, it works just like an
IEnumerable, but is allowed to be async.

So, in order to efficiently process include files, the whole file needs
to be processed at once, while individual tasks write tokens to a
buffer. This was already happening line by line, but the buffers were
being returned in an IAsyncEnumerable. That meant that if one #include
was being read from disk, the next #include couldn't be read until the
first one loaded and control was passed back to the preprocessor.

The solution was to start a task to buffer all the line buffers as
well, and then piece together the buffers as they come in.